requirements.txt
pandas
streamlit
numpy
python-dateutil
datetime
seaborn
matplotlib.pyplot
plotly.express